+++ src/core/configuration.c
@@ -74,7 +74,7 @@ MsQuicConfigurationOpen(
             AllocFailure,
             "Allocation of '%s' failed. (%llu bytes)",
             "QUIC_CONFIGURATION" ,
-            (uint64_t)sizeof(QUIC_CONFIGURATION));
+            sizeof(QUIC_CONFIGURATION));
         Status = QUIC_STATUS_OUT_OF_MEMORY;
         goto Error;
     }
