+++ src/platform/hashtable.c
@@ -485,7 +485,7 @@ Return Value:
                 AllocFailure,
                 "Allocation of '%s' failed. (%llu bytes)",
                 "CXPLAT_HASHTABLE",
-                (uint64_t)sizeof(CXPLAT_HASHTABLE));
+                sizeof(CXPLAT_HASHTABLE));
             return FALSE;
         }
 
@@ -523,7 +523,7 @@ Return Value:
                 AllocFailure,
                 "Allocation of '%s' failed. (%llu bytes)",
                 "second level dir (0)",
-                (uint64_t)(CxPlatComputeSecondLevelDirSize(0) * sizeof(CXPLAT_LIST_ENTRY)));
+                CxPlatComputeSecondLevelDirSize(0) * sizeof(CXPLAT_LIST_ENTRY));
             CxPlatHashtableUninitialize(Table);
             return FALSE;
         }
@@ -563,7 +563,7 @@ Return Value:
                     AllocFailure,
                     "Allocation of '%s' failed. (%llu bytes)",
                     "second level dir (i)",
-                    (uint64_t)(CxPlatComputeSecondLevelDirSize(i) * sizeof(CXPLAT_LIST_ENTRY)));
+                    CxPlatComputeSecondLevelDirSize(i) * sizeof(CXPLAT_LIST_ENTRY));
                 CxPlatHashtableUninitialize(Table);
                 return FALSE;
             }
